requirements.txt
tweepy==4.10.1
textblob==0.17.1
google-cloud-pubsub==2.13.6
google-auth
google-cloud-language
